last minor change

+++ Procedural-Midi/MidiAsset/Source/Midi/Classes/Util/MidiUtil.cpp
@@ -12,7 +12,7 @@ using namespace std;
 * MIDI Unit Conversions
 */
 long MidiUtil::ticksToMs(long ticks, int mpqn, int resolution) {
-	return ((ticks * (uint64)mpqn) / resolution) / 1000;
+	return ((ticks * (unsigned long long)mpqn) / resolution) / 1000;
 }
 long MidiUtil::ticksToMs(long ticks, float bpm, int resolution) {
 	return ticksToMs(ticks, bpmToMpqn(bpm), resolution);
